--- a/Project.toml
+++ b/Project.toml
@@ -6,6 +6,7 @@ desc = " Chain types and utility functions for MCMC simulations."
 version = "0.4.0"
 
 [deps]
+AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
 AbstractMCMC = "80f14c24-f653-4e6a-9b94-39d6b0f70001"
 AxisArrays = "39de3d68-74b9-583c-8d2d-e117c070f3a9"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
@@ -21,6 +22,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [compat]
+AbstractFFTs = "0.5"
 AbstractMCMC = "~0.1"
 AxisArrays = "^0.3"
 DataFrames = "^0.19"
@@ -30,13 +32,13 @@ RecipesBase = "^0.7"
 Showoff = "^0.3"
 SpecialFunctions = "^0.8"
 StatsBase = "^0.32"
-Turing = "^0.7"
 julia = "^1"
 
 [extras]
+FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
 
 [targets]
-test = ["StatsPlots", "Test", "Turing"]
+test = ["FFTW", "StatsPlots", "Test", "Turing"]