--- a/src/MCMCChains.jl
+++ b/src/MCMCChains.jl
@@ -2,6 +2,7 @@ module MCMCChains
 
 using AbstractFFTs
 using AbstractMCMC
+using Compat
 
 import Statistics
 import Showoff: showoff