Sorry for this rather incomplete and unfinished PR, but I realized that I probably won't have time to finish it before leaving to NeurIPS and I guess there are many things to discuss, so I just post a draft, a benchmark, and some comments.

This PR provides 3 different implementations of the effective sample size. The first two (`ESSMethod` and `FFTESSMethod`) are very similar to the implementation in Stan, basically the (only?) difference is that this PR doesn't use Stan's improved estimate for the antithetic case (https://github.com/stan-dev/stan/blob/0281899d26644444384ba11f55fda161d52daa18/src/stan/analyze/mcmc/compute_effective_sample_size.hpp#L119-L122). `ESSMethod` estimates the autocorrelation for each desired lag separately in a sequential manner, whereas `FFTESSMethod` uses a FFT to compute the autocorrelation for all lags "simultaneously" (as Stan does as well). In contrast to these two methods, which are based on the plug-in estimator of the autocorrelation and the discussion in https://arxiv.org/pdf/1903.08008.pdf, the third method `BDAESSMethod` uses the (currently implemented) variogram-based estimation of the autocorrelation, as described, e.g., in the Bayesian Data Analysis book. All methods use the identical way of estimating split-Rhat.

I'm still a bit unsatisfied with the current draft since I think preferably algorithms such as ESS estimation should be implemented in a way that is quite agnostic about the structure of the chains and is not particularly tailored towards `Chains` (as a side remark, I think that holds for most diagnostic tools; it seems a bit unfortunate that, even though implemented for `AbstractChains`, most methods seem to assume a `Chains`-like structure - however, maybe that's an intentional design choice). If the design goal is to have only specific algorithms for `Chains` in this package, maybe generic algorithms should be added to a separate package?

Moreover, in contrast to the current implementation `missing` values lead to `missing` values of the ESS and Rhat. If this should be changed, the implementation has to be modified quite a bit since currently it is assumed and exploited that all chains are of the same length (as in Stan's implementation and most papers I've seen). I doubt that the current implementation handles chains of different sizes (which might result from skipping `missing` values) correctly: the number of samples `N` (in Stan's notation) is just set to the minimum number of samples in a chain for the first parameter (although it could differ for different parameters), and used in the further calculations even though the between-chain and within-chain variances are estimated from all available samples. Intuitively, the calculations should be adjusted and account for the fact that there might be different samples per chain?

The main motivation for this PR was that I wanted to estimate the ESS of a chain of 10000 samples with 128x128 = 16384 parameters (each sample represents a Gaussian random field). I noticed that all Julia implementations that I could find were too slow for my purposes, so I ended up calling a R library from Julia. A simple benchmark shows that this PR leads to an almost 1000-fold speedup compared with the current implementation:
```julia
using MCMCChains
using BenchmarkTools
using FFTW

using Random

Random.seed!(100)

const chain = Chains(randn(10_000, 40, 10))

# Master:

@btime ess($chain)

#=
  10.125 s (4096594 allocations: 43.76 GiB)
ESS

│ Row │ parameters │ ess       │ r_hat    │
│     │ Symbol     │ Any       │ Any      │
├─────┼────────────┼───────────┼──────────┤
│ 1   │ Param1     │ 1.00727e5 │ 1.00001  │
│ 2   │ Param2     │ 97625.1   │ 0.999991 │
│ 3   │ Param3     │ 100563.0  │ 1.0      │
│ 4   │ Param4     │ 97829.3   │ 1.0      │
│ 5   │ Param5     │ 99910.1   │ 1.0      │
│ 6   │ Param6     │ 1.01397e5 │ 1.00004  │
│ 7   │ Param7     │ 1.0016e5  │ 1.00001  │
│ 8   │ Param8     │ 98929.0   │ 1.00008  │
│ 9   │ Param9     │ 1.00422e5 │ 1.00002  │
│ 10  │ Param10    │ 98978.1   │ 0.99997  │
│ 11  │ Param11    │ 99053.8   │ 1.00006  │
│ 12  │ Param12    │ 1.01338e5 │ 1.00005  │
│ 13  │ Param13    │ 1.00735e5 │ 0.999968 │
│ 14  │ Param14    │ 100478.0  │ 1.00001  │
│ 15  │ Param15    │ 101296.0  │ 1.00002  │
│ 16  │ Param16    │ 98307.1   │ 1.00005  │
│ 17  │ Param17    │ 1.00897e5 │ 1.00002  │
│ 18  │ Param18    │ 1.00307e5 │ 0.999981 │
│ 19  │ Param19    │ 1.00508e5 │ 1.00008  │
│ 20  │ Param20    │ 98470.5   │ 1.0      │
│ 21  │ Param21    │ 1.00483e5 │ 0.999983 │
│ 22  │ Param22    │ 99396.9   │ 0.999989 │
│ 23  │ Param23    │ 99395.7   │ 1.00004  │
│ 24  │ Param24    │ 99198.6   │ 1.00002  │
│ 25  │ Param25    │ 1.00437e5 │ 0.99999  │
│ 26  │ Param26    │ 1.0053e5  │ 0.999962 │
│ 27  │ Param27    │ 101602.0  │ 0.999967 │
│ 28  │ Param28    │ 99008.9   │ 0.999951 │
│ 29  │ Param29    │ 1.00593e5 │ 1.00006  │
│ 30  │ Param30    │ 1.01038e5 │ 0.999989 │
│ 31  │ Param31    │ 98576.5   │ 1.00005  │
│ 32  │ Param32    │ 1.01251e5 │ 1.00001  │
│ 33  │ Param33    │ 1.00384e5 │ 1.00001  │
│ 34  │ Param34    │ 1.01561e5 │ 1.00004  │
│ 35  │ Param35    │ 1.00154e5 │ 0.999966 │
│ 36  │ Param36    │ 99093.7   │ 1.0      │
│ 37  │ Param37    │ 1.01628e5 │ 0.999976 │
│ 38  │ Param38    │ 1.00907e5 │ 1.00002  │
│ 39  │ Param39    │ 1.00696e5 │ 0.999991 │
│ 40  │ Param40    │ 99010.8   │ 0.999983 │
=#

# PR:

f(x) = ChainDataFrame("ESS", ess(x), digits=missing)
@btime f($chain)

#=
  14.682 ms (8480 allocations: 1.27 MiB)
ESS

│ Row │ parameters │ ess       │ r_hat    │
│     │ Symbol     │ Float64   │ Float64  │
├─────┼────────────┼───────────┼──────────┤
│ 1   │ Param1     │ 99862.1   │ 1.00001  │
│ 2   │ Param2     │ 97298.1   │ 0.999991 │
│ 3   │ Param3     │ 1.00184e5 │ 1.0      │
│ 4   │ Param4     │ 96887.9   │ 1.0      │
│ 5   │ Param5     │ 99529.4   │ 1.0      │
│ 6   │ Param6     │ 1.00815e5 │ 1.00004  │
│ 7   │ Param7     │ 99460.6   │ 1.00001  │
│ 8   │ Param8     │ 99609.3   │ 1.00008  │
│ 9   │ Param9     │ 1.00632e5 │ 1.00002  │
│ 10  │ Param10    │ 97383.9   │ 0.99997  │
│ 11  │ Param11    │ 98329.0   │ 1.00006  │
│ 12  │ Param12    │ 1.00048e5 │ 1.00005  │
│ 13  │ Param13    │ 1.00753e5 │ 0.999968 │
│ 14  │ Param14    │ 98251.8   │ 1.00001  │
│ 15  │ Param15    │ 1.00393e5 │ 1.00002  │
│ 16  │ Param16    │ 98372.8   │ 1.00005  │
│ 17  │ Param17    │ 99735.0   │ 1.00002  │
│ 18  │ Param18    │ 98832.8   │ 0.999981 │
│ 19  │ Param19    │ 99192.9   │ 1.00008  │
│ 20  │ Param20    │ 98188.7   │ 1.0      │
│ 21  │ Param21    │ 1.00345e5 │ 0.999983 │
│ 22  │ Param22    │ 98523.4   │ 0.999989 │
│ 23  │ Param23    │ 99087.0   │ 1.00004  │
│ 24  │ Param24    │ 98960.4   │ 1.00002  │
│ 25  │ Param25    │ 1.01007e5 │ 0.99999  │
│ 26  │ Param26    │ 1.00183e5 │ 0.999962 │
│ 27  │ Param27    │ 1.00932e5 │ 0.999967 │
│ 28  │ Param28    │ 98818.3   │ 0.999951 │
│ 29  │ Param29    │ 1.0028e5  │ 1.00006  │
│ 30  │ Param30    │ 99613.4   │ 0.999989 │
│ 31  │ Param31    │ 98288.4   │ 1.00005  │
│ 32  │ Param32    │ 99936.2   │ 1.00001  │
│ 33  │ Param33    │ 1.00464e5 │ 1.00001  │
│ 34  │ Param34    │ 1.00673e5 │ 1.00004  │
│ 35  │ Param35    │ 1.00014e5 │ 0.999966 │
│ 36  │ Param36    │ 98879.9   │ 1.0      │
│ 37  │ Param37    │ 1.00067e5 │ 0.999976 │
│ 38  │ Param38    │ 1.00699e5 │ 1.00002  │
│ 39  │ Param39    │ 1.00618e5 │ 0.999991 │
│ 40  │ Param40    │ 99338.2   │ 0.999983 │
=#

g(x) = ChainDataFrame("ESS", ess(x; method = ESSMethod()), digits=missing)
@btime g($chain)

#=
  14.602 ms (8481 allocations: 1.27 MiB)
ESS

│ Row │ parameters │ ess       │ r_hat    │
│     │ Symbol     │ Float64   │ Float64  │
├─────┼────────────┼───────────┼──────────┤
│ 1   │ Param1     │ 99862.1   │ 1.00001  │
│ 2   │ Param2     │ 97298.1   │ 0.999991 │
│ 3   │ Param3     │ 1.00184e5 │ 1.0      │
│ 4   │ Param4     │ 96887.9   │ 1.0      │
│ 5   │ Param5     │ 99529.4   │ 1.0      │
│ 6   │ Param6     │ 1.00815e5 │ 1.00004  │
│ 7   │ Param7     │ 99460.6   │ 1.00001  │
│ 8   │ Param8     │ 99609.3   │ 1.00008  │
│ 9   │ Param9     │ 1.00632e5 │ 1.00002  │
│ 10  │ Param10    │ 97383.9   │ 0.99997  │
│ 11  │ Param11    │ 98329.0   │ 1.00006  │
│ 12  │ Param12    │ 1.00048e5 │ 1.00005  │
│ 13  │ Param13    │ 1.00753e5 │ 0.999968 │
│ 14  │ Param14    │ 98251.8   │ 1.00001  │
│ 15  │ Param15    │ 1.00393e5 │ 1.00002  │
│ 16  │ Param16    │ 98372.8   │ 1.00005  │
│ 17  │ Param17    │ 99735.0   │ 1.00002  │
│ 18  │ Param18    │ 98832.8   │ 0.999981 │
│ 19  │ Param19    │ 99192.9   │ 1.00008  │
│ 20  │ Param20    │ 98188.7   │ 1.0      │
│ 21  │ Param21    │ 1.00345e5 │ 0.999983 │
│ 22  │ Param22    │ 98523.4   │ 0.999989 │
│ 23  │ Param23    │ 99087.0   │ 1.00004  │
│ 24  │ Param24    │ 98960.4   │ 1.00002  │
│ 25  │ Param25    │ 1.01007e5 │ 0.99999  │
│ 26  │ Param26    │ 1.00183e5 │ 0.999962 │
│ 27  │ Param27    │ 1.00932e5 │ 0.999967 │
│ 28  │ Param28    │ 98818.3   │ 0.999951 │
│ 29  │ Param29    │ 1.0028e5  │ 1.00006  │
│ 30  │ Param30    │ 99613.4   │ 0.999989 │
│ 31  │ Param31    │ 98288.4   │ 1.00005  │
│ 32  │ Param32    │ 99936.2   │ 1.00001  │
│ 33  │ Param33    │ 1.00464e5 │ 1.00001  │
│ 34  │ Param34    │ 1.00673e5 │ 1.00004  │
│ 35  │ Param35    │ 1.00014e5 │ 0.999966 │
│ 36  │ Param36    │ 98879.9   │ 1.0      │
│ 37  │ Param37    │ 1.00067e5 │ 0.999976 │
│ 38  │ Param38    │ 1.00699e5 │ 1.00002  │
│ 39  │ Param39    │ 1.00618e5 │ 0.999991 │
│ 40  │ Param40    │ 99338.2   │ 0.999983 │
=#

h(x) = ChainDataFrame("ESS", ess(x; method = FFTESSMethod()), digits=missing)
@btime h($chain)

#=
  181.382 ms (43567 allocations: 7.37 MiB)
ESS

│ Row │ parameters │ ess       │ r_hat    │
│     │ Symbol     │ Float64   │ Float64  │
├─────┼────────────┼───────────┼──────────┤
│ 1   │ Param1     │ 99862.1   │ 1.00001  │
│ 2   │ Param2     │ 97298.1   │ 0.999991 │
│ 3   │ Param3     │ 1.00184e5 │ 1.0      │
│ 4   │ Param4     │ 96887.9   │ 1.0      │
│ 5   │ Param5     │ 99529.4   │ 1.0      │
│ 6   │ Param6     │ 1.00815e5 │ 1.00004  │
│ 7   │ Param7     │ 99460.6   │ 1.00001  │
│ 8   │ Param8     │ 99609.3   │ 1.00008  │
│ 9   │ Param9     │ 1.00632e5 │ 1.00002  │
│ 10  │ Param10    │ 97383.9   │ 0.99997  │
│ 11  │ Param11    │ 98329.0   │ 1.00006  │
│ 12  │ Param12    │ 1.00048e5 │ 1.00005  │
│ 13  │ Param13    │ 1.00753e5 │ 0.999968 │
│ 14  │ Param14    │ 98251.8   │ 1.00001  │
│ 15  │ Param15    │ 1.00393e5 │ 1.00002  │
│ 16  │ Param16    │ 98372.8   │ 1.00005  │
│ 17  │ Param17    │ 99735.0   │ 1.00002  │
│ 18  │ Param18    │ 98832.8   │ 0.999981 │
│ 19  │ Param19    │ 99192.9   │ 1.00008  │
│ 20  │ Param20    │ 98188.7   │ 1.0      │
│ 21  │ Param21    │ 1.00345e5 │ 0.999983 │
│ 22  │ Param22    │ 98523.4   │ 0.999989 │
│ 23  │ Param23    │ 99087.0   │ 1.00004  │
│ 24  │ Param24    │ 98960.4   │ 1.00002  │
│ 25  │ Param25    │ 1.01007e5 │ 0.99999  │
│ 26  │ Param26    │ 1.00183e5 │ 0.999962 │
│ 27  │ Param27    │ 1.00932e5 │ 0.999967 │
│ 28  │ Param28    │ 98818.3   │ 0.999951 │
│ 29  │ Param29    │ 1.0028e5  │ 1.00006  │
│ 30  │ Param30    │ 99613.4   │ 0.999989 │
│ 31  │ Param31    │ 98288.4   │ 1.00005  │
│ 32  │ Param32    │ 99936.2   │ 1.00001  │
│ 33  │ Param33    │ 1.00464e5 │ 1.00001  │
│ 34  │ Param34    │ 1.00673e5 │ 1.00004  │
│ 35  │ Param35    │ 1.00014e5 │ 0.999966 │
│ 36  │ Param36    │ 98879.9   │ 1.0      │
│ 37  │ Param37    │ 1.00067e5 │ 0.999976 │
│ 38  │ Param38    │ 1.00699e5 │ 1.00002  │
│ 39  │ Param39    │ 1.00618e5 │ 0.999991 │
│ 40  │ Param40    │ 99338.2   │ 0.999983 │
=#

k(x) = ChainDataFrame("ESS", ess(x; method = BDAESSMethod()), digits=missing)
@btime k($chain)

#=
  92.891 ms (1121 allocations: 837.20 KiB)
ESS

│ Row │ parameters │ ess       │ r_hat    │
│     │ Symbol     │ Float64   │ Float64  │
├─────┼────────────┼───────────┼──────────┤
│ 1   │ Param1     │ 99833.6   │ 1.00001  │
│ 2   │ Param2     │ 83104.8   │ 0.999991 │
│ 3   │ Param3     │ 1.0015e5  │ 1.0      │
│ 4   │ Param4     │ 96699.2   │ 1.0      │
│ 5   │ Param5     │ 99499.2   │ 1.0      │
│ 6   │ Param6     │ 1.00763e5 │ 1.00004  │
│ 7   │ Param7     │ 99239.0   │ 1.00001  │
│ 8   │ Param8     │ 99143.5   │ 1.00008  │
│ 9   │ Param9     │ 1.006e5   │ 1.00002  │
│ 10  │ Param10    │ 96480.8   │ 0.99997  │
│ 11  │ Param11    │ 96309.0   │ 1.00006  │
│ 12  │ Param12    │ 1.00001e5 │ 1.00005  │
│ 13  │ Param13    │ 1.00709e5 │ 0.999968 │
│ 14  │ Param14    │ 97771.2   │ 1.00001  │
│ 15  │ Param15    │ 1.00359e5 │ 1.00002  │
│ 16  │ Param16    │ 97807.3   │ 1.00005  │
│ 17  │ Param17    │ 99491.8   │ 1.00002  │
│ 18  │ Param18    │ 98031.1   │ 0.999981 │
│ 19  │ Param19    │ 98998.9   │ 1.00008  │
│ 20  │ Param20    │ 97373.5   │ 1.0      │
│ 21  │ Param21    │ 1.00314e5 │ 0.999983 │
│ 22  │ Param22    │ 97216.1   │ 0.999989 │
│ 23  │ Param23    │ 98873.5   │ 1.00004  │
│ 24  │ Param24    │ 98624.4   │ 1.00002  │
│ 25  │ Param25    │ 1.00852e5 │ 0.99999  │
│ 26  │ Param26    │ 1.00071e5 │ 0.999962 │
│ 27  │ Param27    │ 1.0089e5  │ 0.999967 │
│ 28  │ Param28    │ 96266.1   │ 0.999951 │
│ 29  │ Param29    │ 1.00226e5 │ 1.00006  │
│ 30  │ Param30    │ 99570.1   │ 0.999989 │
│ 31  │ Param31    │ 96918.2   │ 1.00005  │
│ 32  │ Param32    │ 99897.0   │ 1.00001  │
│ 33  │ Param33    │ 100426.0  │ 1.00001  │
│ 34  │ Param34    │ 1.00625e5 │ 1.00004  │
│ 35  │ Param35    │ 99981.0   │ 0.999966 │
│ 36  │ Param36    │ 98659.9   │ 1.0      │
│ 37  │ Param37    │ 98804.6   │ 0.999976 │
│ 38  │ Param38    │ 1.00457e5 │ 1.00002  │
│ 39  │ Param39    │ 1.00574e5 │ 0.999991 │
│ 40  │ Param40    │ 99300.0   │ 0.999983 │
=#
```
The benchmarks also show that the Rhat values from the new implementations seem to be identical to the ones in the current implementation. Moreover, the fact that in the current implementation the column types of `ess` and `r_hat` (BTW, maybe it could be called `rhat` instead, consistent with [Julia's naming conventions](https://docs.julialang.org/en/v1/manual/style-guide/index.html#Use-naming-conventions-consistent-with-Julia-base/-1)?) are `Any` already suggests that performance might be suboptimal.

Do you have any suggestions for how to check that the estimates are actually "reasonable"? I chose IID samples here since then I knew what to expect, but maybe we could use some "more interesting" benchmark problems?

Based on the benchmarks, maybe it's even sufficient to implement `ESSMethod`? Of course, that would remove the need for designing an interface for different algorithms. On the other hand, maybe one would like to support even more estimation algorithms? Stan switched completely to an algorithm very similar to `FFTESSMethod`, and interestingly they [state](https://github.com/stan-dev/stan/issues/2450#issue-285135536) that
> Stan reference Section 29.4 describes a variogram-based estimator for
autocorrelations (same as in BDA3). The code uses fast Fourier
transform to compute autocorrelations. Based on the experiments,
computing autocorrelations directly is also more accurate for higher
lags than the variogram estimator given in BDA3. There are other
variogram estimators, but since the current autocorrelations
computation is accurate enough and FFT based computation is faster,
there is no need to look for alternative variogram estimators.

A last comment: due to some additional check for the number of samples (there are many more in Stan it seems) this PR fixes https://github.com/TuringLang/MCMCChains.jl/issues/137 (for one sample it would just return `missing`).