@cpfiffer 2019-12-06
> If the design goal is to have only specific algorithms for `Chains` in this package, maybe generic algorithms should be added to a separate package?

Yeah, I think that's a nice goal. I'm a little unhappy with how `Chains` specific has been -- we could consider moving this stuff upstream into [AbstractMCMC.jl](https://github.com/TuringLang/AbstractMCMC.jl) and providing a very general interface for accessing parameter names and values.

> Intuitively, the calculations should be adjusted and account for the fact that there might be different samples per chain?

So, each parameter should be adjusted by its own minimum across all chains, does that seem right?

> (BTW, maybe it could be called rhat instead, consistent with Julia's naming conventions?)

Sounds good to me, go for it.

> Do you have any suggestions for how to check that the estimates are actually "reasonable"? I chose IID samples here since then I knew what to expect, but maybe we could use some "more interesting" benchmark problems?

I think @goedman and I ran into this earlier when I initially set up the ESS functionality. It's tricky to benchmark and I don't think we came to a general consensus on what works well.

> Based on the benchmarks, maybe it's even sufficient to implement `ESSMethod`? Of course, that would remove the need for designing an interface for different algorithms. On the other hand, maybe one would like to support even more estimation algorithms? 

I think multiple methods is fine -- by default we should just pick one (`ESSMethod`) but allow other methods with keyword arguments. Having the infrastructure to change is also useful if Stan finds some other crazy way of calculating ESS.

I'm quite pleased with this PR, as I was never really happy with the ESS implementation due to speed and general badness, so I'm happy to get the Widmann-womp on this one. I'll give a formal code review once this is out of WIP status.