@goedman 2019-12-06
Hi Cameron,

Since your last updates I've certainly never seen substantial differences (as we saw occasionally earlier).

David's PR and your recent changes are further improvements to the package. Orthogonal to this, maybe at some point in the future a refactoring like what happened to StatsBase, StatsFuns and StatsPlots could help.

Although, as you know Cameron, initially I struggled with the introduction of AxisArrays
but these days I consistently try to use MCMCChains to evaluate the sampling. Any further work I turn the chains into a DataFrame. For me that works fine.

Best,
Rob