@devmotion 2019-12-12
> Yeah, I think that's a nice goal. I'm a little unhappy with how Chains specific has been -- we could consider moving this stuff upstream into AbstractMCMC.jl and providing a very general interface for accessing parameter names and values.

Sounds good, but I guess that should be done in separate PRs.

> So, each parameter should be adjusted by its own minimum across all chains, does that seem right?

Yes, that's what I thought. But it is still not clear to me that just adjusting by the minimum across all chains is the mathematically correct approach. Do you know a reference for the case of chains of different lengths?