@cpfiffer 2019-12-14
I'm not aware of a "real" case, but here's one I cooked up:

```julia
using Turing

@model difflen() = begin
    num_categories ~ Categorical(3)
    draws = Vector(undef, num_categories)
    
    for i in 1:num_categories
        draws[i] ~ Bernoulli(0.5)
    end

    return draws
end

chain = sample(difflen(), PG(100), 1000)
```

The current ESS can't handle this, and throws an out-of-bounds error. 

I remember reading something about this somewhere but I can't seem to find the source -- I'll keep poking around to see if I can turn something up.