@cpfiffer 2019-12-14
Found it. Stan computes ESS by using the minimum number of draws across all chains:

https://github.com/stan-dev/stan/blob/7e5b8882bd3030967138a2102f5fccda6949e654/src/stan/analyze/mcmc/compute_effective_sample_size.hpp#L37

Not sure exactly what the best approach is here to calculate the minimum, if we wanted to hew close to what Stan does. `collect(skipmissing(data))` and then just use the minimum length of those vectors?