@devmotion 2020-01-10
Ah sorry, completely forgot this PR :open_mouth: I'll have a look at it in the next days and try to come up with a final version within the next week!