@devmotion 2020-05-06
Superceded by https://github.com/TuringLang/MCMCChains.jl/pull/215.